pyproject.toml
[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[project]
name = "scanbox"
version = "0.0.1"
description = "Self-contained medical document scanning, splitting, and archival pipeline"
requires-python = ">=3.13"
license = "MIT"
dependencies = [
    "fastapi>=0.135",
    "uvicorn[standard]>=0.34",
    "httpx>=0.28",
    "pikepdf>=10",
    "ocrmypdf>=17",
    "Pillow>=12",
    "pdf2image>=1.17",
    "pytesseract>=0.3",
    "litellm==1.82.6",
    "python-multipart>=0.0.18",
    "aiosqlite>=0.22",
    "jinja2>=3.1",
    "jinja2-fragments>=1.5",
    "mcp>=1.0",
    "zeroconf>=0.140",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=1.3",
    "respx>=0.22",
    "ruff>=0.15",
    "httpx[http2]>=0.28",
    "fpdf2>=2.8",
]

[tool.setuptools.packages.find]
include = ["scanbox*"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
